pyproject.toml
[tool.poetry]
name = "docsearch"
version = "0.1.0"
description = ""
authors = ["bib0x"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
